Fix computed index field on search engine bindings
When assigning first multiple language on the shopinvader backend you
end up creating multiple search engine bindings, one per product and
per lang.
And for each lang that has no index entry in the search engine backend
the field index on binding is empty.  Which is fine.

But, when adding the missing lang in search engine backend indexes,
index field on binding was not recomputed.

The only option to recompute it was to trigger a rewrite on Search
engine backend on the shopinvader backend.

This fix extends the depends on indexes of the binding field index
to be triggered when adding a new index entry in search engine. Thus
you don't need to hack your way anymore with the search engine field
on the shopinvader backend.